requirements.txt
Faker==22.2.0
jsonlines==4.0.0
openai==0.27.0
pandas==2.2.2
Pillow==10.3.0
Requests==2.32.2
retrying==1.3.4
